progress

+++ build_env.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
 
+set -e
+
 #if [[ $# -lt 1 || "$1" == "-h" || "$1" == "--help" || ($# -eq 2 && "$2" != "--upload") || $# -gt 2 ]]; then
 #  echo "Usage: $0 <version>"
 #  exit 1
@@ -15,6 +17,7 @@ i='apt install -y --no-install-recommends'
 
 # essentials
 container=essentials
+img_essentials=aimmspro/devenv-essentials
 
 container_exists(){
   buildah run --runtime /usr/bin/crun $1 -- sh -c ':' 2>/dev/null
@@ -23,7 +26,7 @@ container_exists(){
 
 # shellcheck disable=SC2046
 if [ $(container_exists $container) -ne 0 ]; then
-  buildah from --name essentials ubuntu:$ubuntu_ver
+  buildah from --name $container ubuntu:$ubuntu_ver
   echo "Created new container: $container"
 
   b(){
@@ -51,12 +54,24 @@ if [ $(container_exists $container) -ne 0 ]; then
   b update-alternatives --install /usr/bin/cc cc /usr/bin/clang 100
   b update-alternatives --install /usr/bin/c++ c++ /usr/bin/clang++ 100
   b rm -rf /var/lib/apt/lists/*
+  buildah commit $container $img_essentials
+  buildah rm $container
+fi
 
+container=python
+if [ $(container_exists $container) -ne 0 ]; then
+  buildah from --name $container $img_essentials 
+  echo "Created new container: $container"
+
+  b(){
+    # shellcheck disable=SC2068
+    buildah run --runtime /usr/bin/crun $container -- $@
+  }
   b bash -c 'curl https://pyenv.run | bash'
   b bash -c 'MAKE_OPTS="V=1 -j`grep -c ^processor /proc/cpuinfo`" LLVM_PROFDATA=/usr/bin/llvm-profdata-10 CONFIGURE_OPTS=--enable-optimizations /root/.pyenv/bin/pyenv install $PYTHON_VERSION --verbose'
 
   buildah copy --chown root $container $script_dir/assets/.zshrc /root/.zshrc
-  buildah commit $container aimmspro/devenv-essentials
+  buildah commit $container aimmspro/devenv-python
   buildah rm $container
 fi
